+++ docs/release-notes/change-log.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+* Fixed `--skip-package-validation` to also skip sub packages being imported.
+
 * Trim down packages when using 'imports': only the modules explicitly defined in the YAML manifest and their dependencies will end up in the final spkg.
 
 ## v1.13.0